Refactor warehouse management controllers and update API routes
- Deleted WarehouseCompaniesController and replaced it with dedicated controllers for better organization.
- Added warehouseCompanyListController for listing all companies.
- Added warehouseCompanyShowController for retrieving a specific company by ID.
- Added warehouseCreateProductStockWarehouseController for creating products in stock.
- Added warehouseDeleteCompanyByIdController for deleting a company by ID.
- Added warehouseDeleteStockController for deleting stock items.
- Added warehouseDeleteWarehouseByIdController for deleting a warehouse by ID.
- Added warehouseGetStockByWarehouseController for retrieving stock by warehouse.
- Added warehouseListAllController for listing all warehouses.
- Added warehouseNewCompanyController for creating new companies.
- Added warehouseNewWarehouseStoreController for creating new warehouses.
- Added warehouseShowByIdController for retrieving a specific warehouse by ID.
- Added warehouseUpdateCompanyNameController for updating company names.
- Added warehouseUpdateDetailsController for updating warehouse details.
- Added warehouseUpdateStockForWarehouseController for updating stock for a warehouse.
- Updated API routes to reflect new controller structure and endpoints.
