+++ multistock-backend/app/Http/Controllers/Warehouses/warehouseDeleteWarehouseByIdController.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Http\Controllers\Warehouses;
+
+use App\Models\Company;
+use App\Models\Warehouse;
+use App\Models\StockWarehouse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Http;
+
+class warehouseDeleteWarehouseByIdController{
+    /**
+     * Delete a specific warehouse by ID.
+     */
+    public function warehouse_delete($id)
+    {
+        $warehouse = Warehouse::find($id);
+
+        if ($warehouse) {
+            $warehouse->delete();
+            return response()->json(['message' => 'Bodega eliminada con éxito.']);
+        }
+
+        return response()->json(['message' => 'Bodega no encontrada.'], 404);
+    }
+}
